requirements.txt
scikit-learn
matplotlib
pandas
google.generativeai
PIL
Flask